refactor: Bookmark API 리팩토링
