--- src/main/java/com/flytrap/rssreader/api/bookmark/domain/Bookmark.java
@@ -1,26 +1,0 @@
-package com.flytrap.rssreader.api.bookmark.domain;
-
-import com.flytrap.rssreader.global.model.DefaultDomain;
-import com.flytrap.rssreader.global.model.Domain;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Domain(name = "bookmark")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Bookmark implements DefaultDomain {
-
-    private Long id;
-    private Long memberId;
-    private Long postId;
-
-    @Builder
-    protected Bookmark(Long id, Long memberId, Long postId) {
-        this.id = id;
-        this.memberId = memberId;
-        this.postId = postId;
-    }
-
-}
