--- src/main/java/com/flytrap/rssreader/api/bookmark/presentation/dto/BookmarkRequest.java
@@ -1,13 +1,0 @@
-package com.flytrap.rssreader.api.bookmark.presentation.dto;
-
-import com.flytrap.rssreader.api.bookmark.domain.Bookmark;
-
-public record BookmarkRequest() {
-
-    public record Response(Long id, Long memberId, Long postId) {
-        public static Response from(Bookmark bookmark) {
-            return new Response(bookmark.getId(), bookmark.getMemberId(), bookmark.getPostId());
-        }
-    }
-
-}
