+++ src/main/java/com/flytrap/rssreader/api/post/infrastructure/entity/BookmarkEntity.java
@@ -1,6 +1,5 @@
-package com.flytrap.rssreader.api.bookmark.infrastructure.entity;
+package com.flytrap.rssreader.api.post.infrastructure.entity;
 
-import com.flytrap.rssreader.api.bookmark.domain.Bookmark;
 import com.flytrap.rssreader.api.member.domain.AccountId;
 import com.flytrap.rssreader.api.post.domain.PostId;
 import jakarta.persistence.Column;
@@ -9,7 +8,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import java.util.Objects;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -38,14 +36,6 @@ public class BookmarkEntity {
         this.postId = postId;
     }
 
-    public static BookmarkEntity from(Bookmark bookmark) {
-        return BookmarkEntity.builder()
-            .id(bookmark.getId())
-            .memberId(bookmark.getMemberId())
-            .postId(bookmark.getPostId())
-            .build();
-    }
-
     public static BookmarkEntity create(AccountId accountId, PostId postId) {
         return BookmarkEntity.builder()
             .memberId(accountId.value())
@@ -53,15 +43,4 @@ public class BookmarkEntity {
             .build();
     }
 
-    public Bookmark toDomain() {
-        return Bookmark.builder()
-            .id(id)
-            .memberId(memberId)
-            .postId(postId)
-            .build();
-    }
-
-    public boolean isSameMember(Long memberId) {
-        return Objects.equals(this.memberId, memberId);
-    }
 }
