+++ src/main/java/com/flytrap/rssreader/api/post/infrastructure/implementation/PostUpdater.java
@@ -1,7 +1,7 @@
 package com.flytrap.rssreader.api.post.infrastructure.implementation;
 
-import com.flytrap.rssreader.api.bookmark.infrastructure.entity.BookmarkEntity;
-import com.flytrap.rssreader.api.bookmark.infrastructure.repository.BookmarkEntityJpaRepository;
+import com.flytrap.rssreader.api.post.infrastructure.entity.BookmarkEntity;
+import com.flytrap.rssreader.api.post.infrastructure.repository.BookmarkEntityJpaRepository;
 import com.flytrap.rssreader.api.member.domain.AccountId;
 import com.flytrap.rssreader.api.post.domain.PostAggregate;
 import com.flytrap.rssreader.api.post.infrastructure.entity.OpenEntity;
@@ -10,6 +10,7 @@ import com.flytrap.rssreader.api.post.infrastructure.repository.PostEntityJpaRep
 import com.flytrap.rssreader.api.post.infrastructure.repository.PostOpenEntityRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 @Component
 @RequiredArgsConstructor
@@ -19,7 +20,7 @@ public class PostUpdater {
     private final PostOpenEntityRepository postOpenEntityRepository;
     private final BookmarkEntityJpaRepository bookmarkEntityJpaRepository;
 
-    // Open과 Bookmark 업데이트는 중요하지 않습니다. Post만 업데이트되면 괜찮으므로, @Transactional이 필요하지 않습니다.
+    @Transactional
     public void update(PostAggregate postAggregate, AccountId accountId) {
         PostEntity postEntity = PostEntity.from(postAggregate);
         postEntityJpaRepository.save(postEntity);
@@ -38,7 +39,7 @@ public class PostUpdater {
             .existsByMemberIdAndPostId(accountId.value(), postAggregate.getId().value());
         if (existBookmarkInDB && !postAggregate.isBookmarked()) {
             bookmarkEntityJpaRepository
-                .deleteAllByMemberIdAndPostId(accountId.value(), postAggregate.getId().value());
+                .deleteByMemberIdAndPostId(accountId.value(), postAggregate.getId().value());
         } else if (!existBookmarkInDB && postAggregate.isBookmarked()) {
             BookmarkEntity bookmarkEntity = BookmarkEntity.create(accountId, postAggregate.getId());
             bookmarkEntityJpaRepository.save(bookmarkEntity);
