+++ src/main/java/com/flytrap/rssreader/api/post/infrastructure/repository/BookmarkEntityJpaRepository.java
@@ -1,10 +1,10 @@
-package com.flytrap.rssreader.api.bookmark.infrastructure.repository;
+package com.flytrap.rssreader.api.post.infrastructure.repository;
 
-import com.flytrap.rssreader.api.bookmark.infrastructure.entity.BookmarkEntity;
+import com.flytrap.rssreader.api.post.infrastructure.entity.BookmarkEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookmarkEntityJpaRepository extends JpaRepository<BookmarkEntity, Long> {
     boolean existsByMemberIdAndPostId(Long memberId, Long postId);
 
-    void deleteAllByMemberIdAndPostId(Long memberId, Long postId);
+    void deleteByMemberIdAndPostId(Long memberId, Long postId);
 }
