+++ src/main/java/com/flytrap/rssreader/api/post/infrastructure/repository/PostListReadDslRepository.java
@@ -1,9 +1,9 @@
 package com.flytrap.rssreader.api.post.infrastructure.repository;
 
-import static com.flytrap.rssreader.api.bookmark.infrastructure.entity.QBookmarkEntity.bookmarkEntity;
 import static com.flytrap.rssreader.api.folder.infrastructure.entity.QFolderEntity.folderEntity;
 import static com.flytrap.rssreader.api.folder.infrastructure.entity.QFolderSubscribeEntity.folderSubscribeEntity;
 import static com.flytrap.rssreader.api.folder.infrastructure.entity.QSharedFolderEntity.sharedFolderEntity;
+import static com.flytrap.rssreader.api.post.infrastructure.entity.QBookmarkEntity.bookmarkEntity;
 import static com.flytrap.rssreader.api.post.infrastructure.entity.QOpenEntity.openEntity;
 import static com.flytrap.rssreader.api.post.infrastructure.entity.QPostEntity.postEntity;
 import static com.flytrap.rssreader.api.subscribe.infrastructure.entity.QSubscribeEntity.subscribeEntity;
