+++ src/main/java/com/flytrap/rssreader/api/post/presentation/dto/response/BookmarkResponse.java
@@ -1,0 +1,7 @@
+package com.flytrap.rssreader.api.post.presentation.dto.response;
+
+public record BookmarkResponse(
+    Long accountId, Long postId
+) {
+
+}
